+++ src/main/java/com/construction_worker_forum_back/model/chat/MessageStatus.java
@@ -1,0 +1,5 @@
+package com.construction_worker_forum_back.model.chat;
+
+public enum MessageStatus {
+    RECEIVED, DELIVERED
+}
