--- src/main/java/com/construction_worker_forum_back/model/utils/MessageStatus.java
@@ -1,5 +1,0 @@
-package com.construction_worker_forum_back.model.utils;
-
-public enum MessageStatus {
-    RECEIVED, DELIVERED
-}
